Bug fix and add annotations
Apparently, 'AT' in locus name is not sufficient to identify the last double quoted element. Some lines of the StringTie output are having a gene_name tag as the last double quoted element, when the gene_id tag is not "MSTRG.x" but a locus name already. So some of the gene names has 'AT' in it, and the locus name already in the 'gene_id' tag will be replaced by a gene_name, causing one gene locus becomes two gene entities in the following analysis. 
Therefore, A regular expression is used to match the Arabidopsis locus name, which gives more accurate replacement.